Moved publish or create views to a common namespace
