+++ src/Console/Commands/ConfigBackpackCommand.php
@@ -45,22 +45,11 @@ class ConfigBackpackCommand extends GeneratorCommand
     }
 
     /**
-     * Alias for the fire method.
-     *
-     * In Laravel 5.5 the fire() method has been renamed to handle().
-     * This alias provides support for both Laravel 5.4 and 5.5.
-     */
-    public function handle()
-    {
-        $this->fire();
-    }
-
-    /**
      * Execute the console command.
      *
      * @return bool|null
      */
-    public function fire()
+    public function handle()
     {
         $name = $this->getNameInput();
 
@@ -110,17 +99,5 @@ class ConfigBackpackCommand extends GeneratorCommand
     protected function buildClass($name)
     {
         return $this->files->get($this->getStub());
-    }
-
-    /**
-     * Get the console command options.
-     *
-     * @return array
-     */
-    protected function getOptions()
-    {
-        return [
-
-        ];
     }
 }
