+++ src/Console/Commands/RequestBackpackCommand.php
@@ -54,16 +54,4 @@ class RequestBackpackCommand extends GeneratorCommand
     {
         return $rootNamespace.'\Http\Requests';
     }
-
-    /**
-     * Get the console command options.
-     *
-     * @return array
-     */
-    protected function getOptions()
-    {
-        return [
-
-        ];
-    }
 }
