+++ src/Console/Commands/ViewBackpackCommand.php
@@ -49,22 +49,11 @@ class ViewBackpackCommand extends GeneratorCommand
     }
 
     /**
-     * Alias for the fire method.
-     *
-     * In Laravel 5.5 the fire() method has been renamed to handle().
-     * This alias provides support for both Laravel 5.4 and 5.5.
-     */
-    public function handle()
-    {
-        $this->fire();
-    }
-
-    /**
      * Execute the console command.
      *
      * @return bool|null
      */
-    public function fire()
+    public function handle()
     {
         $name = $this->getNameInput();
 
@@ -114,17 +103,5 @@ class ViewBackpackCommand extends GeneratorCommand
     protected function buildClass($name)
     {
         return $this->files->get($this->getStub());
-    }
-
-    /**
-     * Get the console command options.
-     *
-     * @return array
-     */
-    protected function getOptions()
-    {
-        return [
-
-        ];
     }
 }
